Extract avatar SDK surfaces from world API (#34)
Introduce freven_avatar_sdk_types and freven_avatar_api as the public home for avatar identity, control, controller, lifecycle, and presentation contracts. Move avatar-facing client and controller surfaces out of freven_world_api, keep world ownership boundaries explicit, and add hidden registration hooks so avatar-owned extensions can register through ModContext without leaving those contracts world-owned.
