+++ Cargo.toml
@@ -1,6 +1,8 @@
 [workspace]
 resolver = "2"
 members = [
+    "crates/freven_avatar_sdk_types",
+    "crates/freven_avatar_api",
     "crates/freven_mod_api",
     "crates/freven_guest",
     "crates/freven_guest_sdk",
@@ -25,10 +27,13 @@ documentation = "https://github.com/frevenengine/freven-sdk/tree/main/docs"
 publish = false
 
 [workspace.dependencies]
+freven_avatar_sdk_types = { path = "crates/freven_avatar_sdk_types", version = "0.1.0" }
+freven_avatar_api = { path = "crates/freven_avatar_api", version = "0.1.0" }
 freven_mod_api = { path = "crates/freven_mod_api", version = "0.1.0" }
 freven_guest = { path = "crates/freven_guest", version = "0.1.0" }
 freven_guest_sdk = { path = "crates/freven_guest_sdk", version = "0.1.0" }
 freven_sdk_types = { path = "crates/freven_sdk_types", version = "0.1.0" }
+freven_world_api = { path = "crates/freven_world_api", version = "0.1.0" }
 freven_world_guest = { path = "crates/freven_world_guest", version = "0.1.0" }
 freven_world_sdk_types = { path = "crates/freven_world_sdk_types", version = "0.1.0" }
 freven_volumetric_sdk_types = { path = "crates/freven_volumetric_sdk_types", version = "0.1.0" }
