+++ crates/freven_avatar_api/src/identity.rs
@@ -1,0 +1,1 @@
+pub use freven_avatar_sdk_types::identity::*;
