+++ crates/freven_avatar_sdk_types/src/identity.rs
@@ -1,0 +1,2 @@
+/// Stable avatar/player identity used by avatar-owned public contracts.
+pub type PlayerId = u64;
